unificado request de empleado.
por alguna razon tuve un lapso mental y hice dos request de empleado, uno para crear y otro para actualizar, cuando de hecho con una sola clase de request bastaba
